Use ZonedDateTime for attemptedQueryTime
