+++ supporter-product-data/src/main/scala/com/gu/lambdas/QueryZuoraLambda.scala
@@ -1,15 +1,14 @@
 package com.gu.lambdas
 
 import com.amazonaws.services.lambda.runtime.Context
-import com.gu.conf.ZuoraQuerierConfig
 import com.gu.lambdas.QueryZuoraLambda.queryZuora
 import com.gu.model.Stage
 import com.gu.model.states.{FetchResultsState, QueryType, QueryZuoraState}
 import com.gu.okhttp.RequestRunners.configurableFutureRunner
-import com.gu.services.ZuoraQuerierService
+import com.gu.services.{ConfigService, ZuoraQuerierService}
 import com.typesafe.scalalogging.StrictLogging
 
-import java.time.LocalDateTime
+import java.time.{ZoneId, ZonedDateTime}
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration._
 
@@ -23,9 +22,13 @@ class QueryZuoraLambda extends Handler[QueryZuoraState, FetchResultsState] {
 object QueryZuoraLambda extends StrictLogging {
   def queryZuora(stage: Stage, queryType: QueryType) = {
     logger.info(s"Attempting to submit ${queryType.value} query to Zuora")
-    val attemptedQueryTime = LocalDateTime.now.minusMinutes(1)
+
+    // Get the time we started the query. Docs for why we need to do this are here:
+    // https://knowledgecenter.zuora.com/Central_Platform/API/AB_Aggregate_Query_API/B_Submit_Query/e_Post_Query_with_Retrieval_Time
+    val attemptedQueryTime = ZonedDateTime.now(ZoneId.of("America/Los_Angeles")).minusMinutes(1)
+
     for {
-      config <- ZuoraQuerierConfig.load(stage)
+      config <- ConfigService(stage).load
       service = new ZuoraQuerierService(config, configurableFutureRunner(60.seconds))
       result <- service.postQuery(queryType)
     } yield {
