+++ supporter-product-data/src/main/scala/com/gu/model/states/UpdateDynamoState.scala
@@ -3,13 +3,13 @@ package com.gu.model.states
 import io.circe.generic.semiauto.{deriveDecoder, deriveEncoder}
 import io.circe.{Decoder, Encoder}
 
-import java.time.LocalDateTime
+import java.time.ZonedDateTime
 
 case class UpdateDynamoState(
   filename: String,
   recordCount: Int,
   processedCount: Int,
-  attemptedQueryTime: LocalDateTime
+  attemptedQueryTime: ZonedDateTime
 )
 
 object UpdateDynamoState{
